+++ trunk/Risicammara-java/src/risicammaraServer/messaggiManage/MessaggioAttaccoVinto.java
@@ -43,7 +43,7 @@ public class MessaggioAttaccoVinto implements Messaggio{
     }
 
     @Override
-    public int getSender() {
+    public long getSender() {
         return -1;
     }
